requirements_temp.txt
abnf==2.2.0
accelerate==1.7.0
aioboto3==15.0.0
aiobotocore==2.23.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioitertools==0.12.0
aiosignal==1.4.0
airportsdata==20250706
annotated-types==0.7.0
anyio==4.9.0
astor==0.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
awscli==1.40.26
backoff==2.2.1
bitsandbytes==0.46.1
blake3==1.0.5
boto3==1.38.27
botocore==1.38.27
cachetools==6.1.0
certifi==2025.7.14
cffi==2.0.0b1
chardet==5.2.0
charset-normalizer==3.4.2
cint==1.0.0
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.3
compressed-tensors==0.9.3
contourpy==1.3.2
cryptography==45.0.5
cupy-cuda12x==13.5.1
cut-cross-entropy==25.1.1
cycler==0.12.1
datasets==4.0.0
debugpy==1.8.15
decorator==5.2.1
deprecated==1.2.18
depyf==0.18.0
diffusers==0.34.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docstring-parser==0.17.0
docutils==0.19
einops==0.8.1
email-validator==2.2.0
eval-type-backport==0.2.2
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
fastrlock==0.8.3
fickling==0.1.4
filelock==3.18.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.3.0
gguf==0.17.1
gitdb==4.0.12
gitpython==3.1.45
googleapis-common-protos==1.70.0
gql==3.6.0b4
graphql-core==3.3.0a9
graphviz==0.21
grpcio==1.74.0
h11==0.16.0
hf-transfer==0.1.9
hf-xet==1.1.6rc2
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.0rc0
idna==3.10
importlib-metadata==8.0.0
interegular==0.3.3
intervaltree==3.1.0
ipykernel==6.30.0
ipython==8.37.0
ipywidgets==8.1.7
jedi==0.19.2
jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-client==8.6.3
jupyter-core==5.8.1
jupyterlab-widgets==3.0.15
kaitaistruct==0.10
kiwisolver==1.4.8
lark==1.2.2
levenshtein==0.27.1
litellm==1.74.9.post1
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.10.11
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral-common==1.8.3
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.3
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.97.1
opencv-python-headless==4.12.0.88
openpipe-art==0.3.11.post2
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-semantic-conventions-ai==0.4.12
outlines==0.1.11
outlines-core==0.1.26
packaging==25.0
pandas==2.3.1
panza==0.1.0
parso==0.8.4
partial-json-parser==0.2.1.1.post6
pdfminer-six==20240706
peft==0.16.0
pexpect==4.9.0
pillow==11.3.0
pip==22.0.2
platformdirs==4.3.8
polars==1.32.0b1
polyfile-weave==0.5.6
prometheus-client==0.22.1
prometheus-fastapi-instrumentator==7.1.0
prompt-toolkit==3.0.51
propcache==0.3.2
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pycountry==24.6.1
pycparser==2.22
pydantic==2.12.0a1
pydantic-core==2.37.2
pydantic-extra-types==2.10.5
pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==4.0.0.dev0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
pyzmq==27.0.0
rapidfuzz==3.13.0
ray==2.48.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.14.9
rignore==0.6.4
rpds-py==0.26.0
rsa==4.7.2
s3transfer==0.13.1
safetensors==0.6.0rc0
scipy==1.15.3
seaborn==0.13.2
semver==3.0.4
sentencepiece==0.2.0
sentry-sdk==2.34.0
setproctitle==1.3.6
setuptools==80.9.0
shellingham==1.5.4
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
starlette==0.47.2
stdlib-list==0.11.1
sympy==1.13.1
tblib==3.1.0
tenacity==9.1.2
tiktoken==0.9.0
tokenizers==0.21.4
torch==2.6.0
torchao==0.12.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
triton==3.2.0
trl==0.15.2
typeguard==4.4.4
typer==0.16.0
typing-extensions==4.14.1
typing-inspection==0.4.1
tyro==0.9.26
tzdata==2025.2
unsloth==2025.5.1
unsloth-zoo==2025.5.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
vllm==0.8.5.post1
wandb==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
weave==0.51.59
websockets==15.0.1
wheel==0.45.1
widgetsnbextension==4.0.14
wrapt==1.17.2
xformers==0.0.29.post2
xgrammar==0.1.18
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
